third_party/blink/web_tests/fast/multicol/span/offset-properties-empty-content.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div style="position:relative; border:10px solid; columns:3; gap:0; width:300px;">
    <div></div>
    <!-- In front of the spanner there's only a zero-height DIV. This does still count as column
         content, and we'll therefore establish a zero-height column set in front of the spanner, as
         well as one after it (which happens to have non-zero-height content). No flow thread offset
         should map to the first column set, since it's completely empty. -->
    <div style="column-span:all; height:50px;"></div>
    <div id="elm1" style="height:100px;"></div>
    <div id="elm2" style="height:100px;"></div>
    <div id="elm3" style="height:100px;"></div>
</div>
<script>
    test(function() {
        assert_equals(document.getElementById('elm1').offsetTop, 50);
        assert_equals(document.getElementById('elm2').offsetTop, 50);
        assert_equals(document.getElementById('elm3').offsetTop, 50);
    }, "columns following a spanner following empty content");
</script>
